requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.1
openai>=1.58.1
pinecone-client==5.0.1
langgraph>=1.0.0
langchain-core>=1.0.0
langchain-openai>=1.0.0
langchain-text-splitters>=1.0.0
pdfplumber==0.11.4
python-dotenv==1.0.1
httpx==0.28.1
psycopg2-binary==2.9.10
